Refactored Boltzmann example docs and scripts.
Made substantial edits to the Boltzmann wealth distribution example. The
general train of thought wasn't changed in the .md file, but many
paragraph were restructured significantly and a few paragraphs were
added.

Also, the corresponding files in `examples/` were edited as well. Since
the docs page gives two examples of the Boltzmann model (with a grid and
without), two separate `.jl` files where created so that either
variation can be easily executed by the user. The code referenced in
docs page is exactly the same code in the two `.jl` files.