Improve documentation of Pathfinder
@AayushSabharwal please have a look at my changes. Documentation should be harmonious in the way it works. Further more, it is better to state default values with equals (`diagonal_movement = true`) instead of saying their default values by English words.

The syntax you use, that allows users to pass both DistanceMetric, and DistanceMetric(), is a bad idea. In Julia these are two different things, and should be treated as such. The "convenience" of not having to type `()` is a null argument. It makes no sense. Let's keep things and enforce that input arguments are what they should be.